services/event-logger/requirements.txt
redis
python-dotenv
psycopg2-binary